+++ buildSrc/src/Extensions.kt
@@ -2,7 +2,6 @@ package com.serebit.strife.buildsrc
 
 import groovy.util.Node
 import org.gradle.api.Project
-import org.gradle.api.artifacts.Dependency
 import org.gradle.api.artifacts.dsl.RepositoryHandler
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 import org.gradle.api.publish.PublishingExtension
@@ -11,26 +10,35 @@ import org.gradle.api.tasks.bundling.Jar
 import org.gradle.kotlin.dsl.maven
 import org.jetbrains.kotlin.gradle.plugin.KotlinDependencyHandler
 
-fun KotlinDependencyHandler.kotlinx(module: String, version: String): String = "org.jetbrains.kotlinx:kotlinx-$module:$version"
-fun KotlinDependencyHandler.ktor(module: String, version: String): String = "io.ktor:ktor-$module:$version"
-fun KotlinDependencyHandler.api(group: String, name: String, version: String): Dependency? = api("$group:$name:$version")
-fun KotlinDependencyHandler.implementation(group: String, name: String, version: String): Dependency? =
+fun KotlinDependencyHandler.kotlinx(module: String, version: String) = "org.jetbrains.kotlinx:kotlinx-$module:$version"
+fun KotlinDependencyHandler.ktor(module: String, version: String) = "io.ktor:ktor-$module:$version"
+fun KotlinDependencyHandler.api(group: String, name: String, version: String) = api("$group:$name:$version")
+fun KotlinDependencyHandler.implementation(group: String, name: String, version: String) =
     implementation("$group:$name:$version")
 
-fun RepositoryHandler.kotlinx(): MavenArtifactRepository = maven("https://kotlin.bintray.com/kotlinx")
-fun RepositoryHandler.kotlinEap(): MavenArtifactRepository = maven("https://kotlin.bintray.com/kotlin-eap")
+fun RepositoryHandler.kotlinx() = maven("https://kotlin.bintray.com/kotlinx")
+fun RepositoryHandler.kotlinEap() = maven("https://kotlin.bintray.com/kotlin-eap")
 
-val Project.fullPath: String get() = "${ProjectInfo.name}${project.path.replace(":", "-")}"
+val Project.fullPath get() = "${ProjectInfo.name}${project.path.replace(":", "-")}"
 
-fun PublishingExtension.createBintrayRepository(accessKey: String?): MavenArtifactRepository =
-    repositories.maven("https://api.bintray.com/maven/serebit/public/${ProjectInfo.name}/;publish=0") {
-        name = "bintray"
-
-        credentials {
-            username = "serebit"
-            accessKey?.let { password = it }
-        }
+fun PublishingExtension.createBintrayRepositories() {
+    fun MavenArtifactRepository.applyCredentials() = credentials {
+        username = "serebit"
+        System.getenv("BINTRAY_KEY")?.let { password = it }
     }
+
+    // create public
+    repositories.maven("https://api.bintray.com/maven/serebit/public/${ProjectInfo.name}/;publish=0") {
+        name = "public"
+        applyCredentials()
+    }
+
+    // create snapshot
+    repositories.maven("https://api.bintray.com/maven/serebit/snapshot/${ProjectInfo.name}/;publish=1") {
+        name = "snapshot"
+        applyCredentials()
+    }
+}
 
 private fun Node.add(key: String, value: String) = appendNode(key).setValue(value)
 private inline fun Node.node(key: String, content: Node.() -> Unit) = appendNode(key).also(content)