+++ buildSrc/src/Versions.kt
@@ -2,16 +2,10 @@ package com.serebit.strife.buildsrc
 
 /** Versions of dependencies for type-safe consistency. */
 object Versions {
-    /** Kotlinx Coroutines. */
     const val COROUTINES: String = "1.3.2"
-    /** Ktor */
     const val KTOR: String = "1.2.4"
-    /** Kotlinx Serialization */
     const val SERIALIZATION: String = "0.13.0"
-    /** Logger */
     const val LOGKAT: String = "0.4.7"
-    /** Soywiz Klock */
     const val KLOCK: String = "1.7.3"
-    /** JUnit Jupiter */
     const val JUPITER: String = "5.5.2"
 }